+++ lib/src/spec_context.dart
@@ -2,7 +2,7 @@ part of softhai.spec_dart;
 
 // Public interface
 abstract class SpecContext {
-  static OutputFormatter output = new ConsoleOutputFormatter();
+  static OutputFormatter output = new TextOutputFormatter();
   static String indent = "  ";
   static SpecLanguage language = SpecLanguage.en;